Split core model boundaries
