+++ Sources/QuillCodeCore/LocalEnvironmentAction.swift
@@ -1,0 +1,35 @@
+import Foundation
+
+public struct LocalEnvironmentAction: Codable, Sendable, Hashable, Identifiable {
+    public var id: String
+    public var title: String
+    public var detail: String?
+    public var relativePath: String
+    public var command: String
+    public var sortOrder: Int?
+    public var environment: [String: String]?
+    public var workingDirectory: String?
+    public var timeoutSeconds: Int?
+
+    public init(
+        id: String,
+        title: String,
+        detail: String? = nil,
+        relativePath: String,
+        command: String,
+        sortOrder: Int? = nil,
+        environment: [String: String]? = nil,
+        workingDirectory: String? = nil,
+        timeoutSeconds: Int? = nil
+    ) {
+        self.id = id
+        self.title = title
+        self.detail = detail
+        self.relativePath = relativePath
+        self.command = command
+        self.sortOrder = sortOrder
+        self.environment = environment
+        self.workingDirectory = workingDirectory
+        self.timeoutSeconds = timeoutSeconds
+    }
+}
