+++ Tests/QuillCodeParityTests/ParityWorkspaceProjectGateTests.swift
@@ -74,15 +74,15 @@ final class ParityWorkspaceProjectGateTests: QuillCodeParityTestCase {
 
     func testProjectInstructionScopesStayInCorePromptAndActivityContracts() throws {
         let modelText = try Self.appSourceText(named: "WorkspaceModel.swift")
-        let projectModelsText = try Self.coreSourceText(named: "ProjectModels.swift")
+        let projectInstructionText = try Self.coreSourceText(named: "ProjectInstruction.swift")
         let loaderText = try Self.appSourceText(named: "ProjectInstructionLoader.swift")
         let promptText = try Self.agentSourceText(named: "TrustedRouterPromptBuilder.swift")
         let activityText = try Self.appSourceText(named: "WorkspaceActivitySourceSurfaceBuilder.swift")
         let diagnosticsText = try Self.appSourceText(named: "ProjectInstructionDiagnosticsBuilder.swift")
 
-        XCTAssertTrue(projectModelsText.contains("public var scopePath"), "Instruction applicability should be part of the shared instruction model.")
-        XCTAssertTrue(projectModelsText.contains("static func scopePath(for instructionPath"), "Scope derivation should be centralized in the shared model.")
-        XCTAssertTrue(projectModelsText.contains("static func scopeLabel(for scopePath"), "Scope display labels should be centralized in the shared model.")
+        XCTAssertTrue(projectInstructionText.contains("public var scopePath"), "Instruction applicability should be part of the shared instruction model.")
+        XCTAssertTrue(projectInstructionText.contains("static func scopePath(for instructionPath"), "Scope derivation should be centralized in the shared model.")
+        XCTAssertTrue(projectInstructionText.contains("static func scopeLabel(for scopePath"), "Scope display labels should be centralized in the shared model.")
         XCTAssertTrue(loaderText.contains("ProjectInstruction.scopePath(for: relativePath)"), "Local instruction loading should persist explicit applicability scope.")
         XCTAssertTrue(promptText.contains("Scope: \\(instruction.scopeLabel)"), "TrustedRouter prompt context should expose every instruction block scope.")
         XCTAssertTrue(promptText.contains("Apply whole-project instructions everywhere"), "Prompt policy should distinguish project-wide and subtree instructions.")
