+++ docs/DECISIONS.md
@@ -108,9 +108,9 @@
 - Core domain models live in focused files rather than a general `Models.swift` bucket. `AgentMode.swift`, `ChatModels.swift`, `AgentPlanModels.swift`, `SubagentModels.swift`, `ApprovalModels.swift`, `ThreadEventModels.swift`, `MemoryModels.swift`, `ChatThread.swift`, and `JSONHelpers.swift` own the general thread/chat/memory domain; parity gates prevent the old umbrella file from returning.
 - TrustedRouter model defaults and catalog normalization live in focused core files. `TrustedRouterDefaults.swift` owns provider IDs, aliases, bundled fallback rows, Nike 1.0/Synth branding, and sort/category policy; `ModelInfo.swift` owns the catalog value records.
 - App configuration is a focused core boundary. `AppConfig.swift` owns persisted model/mode/base-URL settings, OAuth/developer-override compatibility, signed-in TrustedRouter account metadata, and favorite-model normalization; general chat/thread/project models should not own settings compatibility rules.
-- Core tool records are a focused schema boundary. `ToolModels.swift` owns `ToolDefinition`, `ToolCall`, `ToolResult`, built-in core tool definitions, tool-call transcript redaction, and tool-specific browser/memory output compatibility; general chat/thread/project models should not own tool payload compatibility rules.
+- Core tool records are a focused schema boundary made of single-purpose files. `ToolDefinition.swift`, `ToolCall.swift`, `ToolResult.swift`, `CoreToolDefinitions.swift`, `BrowserInspectionToolOutput.swift`, and `MemoryRememberToolOutput.swift` own tool schema records, redaction, built-in tool definitions, and tool-specific output compatibility; general chat/thread/project models and broad catch-all tool files should not own tool payload compatibility rules.
 - Core automation records are a focused scheduling boundary. `AutomationModels.swift` owns automation kind/status/schedule enums, recurrence interval semantics, next-run calculation, and display sorting; general chat/thread/project models should not own recurring-workflow scheduling rules.
-- Core project records are a focused workspace boundary. `ProjectModels.swift` owns local/SSH connection parsing and display, project refs, instructions, local environment actions, and project extension manifests; general thread/message models should not own workspace connection or extension compatibility.
+- Core project records are a focused workspace boundary made of single-purpose files. `ProjectConnection.swift`, `ProjectRef.swift`, `ProjectInstruction.swift`, `LocalEnvironmentAction.swift`, and `ProjectExtensionManifest.swift` own local/SSH connection parsing and display, project refs, instructions, local environment actions, and project extension manifests; general thread/message models and broad catch-all project files should not own workspace connection or extension compatibility.
 - Tool-router feature families should delegate before they sprawl. `ToolRouter` owns the common shell/file/patch entry point and composes git definitions from `GitToolCallDispatcher`; local git, GitHub PR, and worktree tool-call argument mapping lives in that dispatcher. This keeps the shared router small while preserving one public `ToolRouter.execute` API for the agent runtime.
 - Shell request policy is a tool-family concern. `ShellToolCallDispatcher` owns `host.shell.run` definitions, cwd containment, timeout bounds, environment override validation, and `ShellExecutionRequest` construction; `ToolRouter` only delegates shell calls before handling file and patch primitives.
 - SSH Remote project integration coverage lives in `WorkspaceRemoteProjectIntegrationTests`, not `WorkspaceModelTests`. SSH project creation, remote context refresh, remote-safe tool exposure, shell/file/git/GitHub/worktree execution through SSH, and remote path safety cross the workspace model, SSH executor, tool cards, surfaces, and transcript events; keep those flows together while low-level remote command builders stay in their focused tests.
